portfolio_be/requirements.txt
flask
flask-cors
flask-sqlalchemy
pymysql
pytest
